+++ WithoutHaste.Sequences/WithoutHaste.Sequences/EmirpPrime.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 
 namespace WithoutHaste.Sequences
 {
@@ -11,21 +12,21 @@ namespace WithoutHaste.Sequences
 	public class EmirpPrime : Sequence
 	{
 		/// <inheritdoc/>
-		public override long[] Numbers { get { return (long[])numbers.Clone(); } }
-		private long[] numbers;
+		public override BigInteger[] Numbers { get { return (BigInteger[])numbers.Clone(); } }
+		private BigInteger[] numbers;
 
 		/// <inheritdoc/>
-		internal static new long[] TestNumbers = new long[] { 13, 17, 31, 37, 71, 73, 79, 97, 107, 113, 149, 157, 167, 179, 199, 311, 337, 347, 359, 389, 701, 709, 733, 739, 743, 751, 761, 769, 907, 937, 941, 953, 967, 971, 983, 991 };
+		internal static new BigInteger[] TestNumbers = new BigInteger[] { 13, 17, 31, 37, 71, 73, 79, 97, 107, 113, 149, 157, 167, 179, 199, 311, 337, 347, 359, 389, 701, 709, 733, 739, 743, 751, 761, 769, 907, 937, 941, 953, 967, 971, 983, 991 };
 
-		public EmirpPrime(long max) : base(max)
+		public EmirpPrime(BigInteger max) : base(max)
 		{
 		}
 
 		protected override void Generate()
 		{
-			long[] primes = new Prime(Max * 10).Numbers; //going order of magnitude higher than prime because reversing prime could end up with a higher number; ex: 17 => 71
-			List<long> emirpPrimes = new List<long>();
-			foreach(long p in primes.Where(p => p <= Max))
+			BigInteger[] primes = new Prime(Max * 10).Numbers; //going order of magnitude higher than prime because reversing prime could end up with a higher number; ex: 17 => 71
+			List<BigInteger> emirpPrimes = new List<BigInteger>();
+			foreach(BigInteger p in primes.Where(p => p <= Max))
 			{
 				if(IsEmirpPrime(p, primes))
 					emirpPrimes.Add(p);
@@ -33,10 +34,10 @@ namespace WithoutHaste.Sequences
 			numbers = emirpPrimes.ToArray();
 		}
 
-		private bool IsEmirpPrime(long p, long[] primes)
+		private bool IsEmirpPrime(BigInteger p, BigInteger[] primes)
 		{
 			string s = String.Join("", p.ToString().ToArray().Select(c => c.ToString()).Reverse().ToArray());
-			long p2 = Int64.Parse(s);
+			BigInteger p2 = BigInteger.Parse(s);
 			if(p2 == p)
 				return false;
 			return (primes.Contains(p2));
