+++ WithoutHaste.Sequences/WithoutHaste.Sequences/HiggsPrime.cs
@@ -17,26 +17,26 @@ namespace WithoutHaste.Sequences
 	public class HiggsPrime : Sequence
 	{
 		/// <inheritdoc/>
-		public override long[] Numbers { get { return (long[])numbers.Clone(); } }
-		private long[] numbers;
+		public override BigInteger[] Numbers { get { return (BigInteger[])numbers.Clone(); } }
+		private BigInteger[] numbers;
 
 		/// <inheritdoc/>
-		internal static new long[] TestNumbers = new long[] { 2, 3, 5, 7, 11, 13, 19, 23, 29, 31, 37, 43, 47, 53, 59, 61, 67, 71, 79, 101, 107, 127, 131, 139, 149, 151, 157, 173, 181, 191, 197, 199, 211, 223, 229, 263, 269, 277, 283, 311, 317, 331, 347, 349 };
+		internal static new BigInteger[] TestNumbers = new BigInteger[] { 2, 3, 5, 7, 11, 13, 19, 23, 29, 31, 37, 43, 47, 53, 59, 61, 67, 71, 79, 101, 107, 127, 131, 139, 149, 151, 157, 173, 181, 191, 197, 199, 211, 223, 229, 263, 269, 277, 283, 311, 317, 331, 347, 349 };
 
-		public HiggsPrime(long max) : base(max)
+		public HiggsPrime(BigInteger max) : base(max)
 		{
 		}
 
 		protected override void Generate()
 		{
 			// Primes p for which p−1 divides the square of the product of all earlier terms
-			long[] primes = new Prime(Max).Numbers;
-			List<long> higgsPrimes = new List<long>();
-			long productOfHiggsPrimes = 1;
-			foreach(long p in primes)
+			BigInteger[] primes = new Prime(Max).Numbers;
+			List<BigInteger> higgsPrimes = new List<BigInteger>();
+			BigInteger productOfHiggsPrimes = 1;
+			foreach(BigInteger p in primes)
 			{
-				long totient = p - 1;
-				long squaredProduct = productOfHiggsPrimes * productOfHiggsPrimes;
+				BigInteger totient = p - 1;
+				BigInteger squaredProduct = productOfHiggsPrimes * productOfHiggsPrimes;
 				if(squaredProduct % totient == 0)
 				{
 					higgsPrimes.Add(p);
